Consolidate FieldService into DotFieldService
Removes the local `FieldService` from the content-types-edit portlet and merges its functionality into the shared `DotFieldService` in `@dotcms/data-access`. Moves the `FieldType` interface to `@dotcms/dotcms-models` as the canonical definition. Scopes `DotEditContentTypeCacheService` to the dialog instance instead of app-wide singleton, and extracts `getFieldIcon` as a pure utility function.
