+++ core-web/apps/dotcms-ui/src/app/portlets/shared/dot-content-types-edit/components/fields/content-type-fields-properties-form/field-properties/dot-relationships-property/services/dot-edit-content-type-cache.service.ts
@@ -8,7 +8,7 @@ import { DotCMSContentType } from '@dotcms/dotcms-models';
  * @export
  * @class DotEditContentTypeCacheService
  */
-@Injectable({ providedIn: 'root' })
+@Injectable()
 export class DotEditContentTypeCacheService {
     private currentContentType: DotCMSContentType;
