+++ core-web/apps/dotcms-ui/src/app/portlets/shared/dot-content-types-edit/components/fields/content-types-fields-list/content-types-fields-list.component.ts
@@ -4,13 +4,13 @@ import { Component, inject, Input, OnInit, signal } from '@angular/core';
 
 import { filter, mergeMap, take, toArray } from 'rxjs/operators';
 
+import { DotFieldService } from '@dotcms/data-access';
 import { DotCMSClazz, DotCMSClazzes } from '@dotcms/dotcms-models';
 import { FieldUtil } from '@dotcms/utils';
 
 import { FIELD_ICONS } from './content-types-fields-icon-map';
 
 import { FieldType } from '..';
-import { FieldService } from '../service';
 /**
  * Show all the Field Types
  *
@@ -45,7 +45,7 @@ export class ContentTypesFieldsListComponent implements OnInit {
 
     #backListFields = ['relationships_tab', 'permissions_tab', 'tab_divider'];
 
-    readonly #fieldService = inject(FieldService);
+    readonly #fieldService = inject(DotFieldService);
 
     ngOnInit(): void {
         this.#fieldService
