+++ core-web/apps/dotcms-ui/src/app/portlets/shared/dot-content-types-edit/components/fields/models/field-type.model.ts
@@ -1,7 +1,3 @@
-export interface FieldType {
-    id: string;
-    label: string;
-    clazz: string;
-    helpText: string;
-    properties: string[];
-}
+// Re-exported from the shared models lib so feature code can keep importing `FieldType`
+// via `../models` while the canonical definition lives in `@dotcms/dotcms-models`.
+export { FieldType } from '@dotcms/dotcms-models';
