+++ core-web/apps/dotcms-ui/src/app/portlets/shared/dot-content-types-edit/components/fields/service/index.ts
@@ -1,3 +1,2 @@
 export * from './field-drag-drop.service';
 export * from './field-properties.service';
-export * from './field.service';
