+++ core-web/libs/data-access/src/lib/dot-field/dot-field.service.ts
@@ -5,7 +5,13 @@ import { inject, Injectable } from '@angular/core';
 
 import { map } from 'rxjs/operators';
 
-import { DotCMSAPIResponse, DotCMSContentTypeField } from '@dotcms/dotcms-models';
+import {
+    DotCMSAPIResponse,
+    DotCMSContentTypeField,
+    DotCMSContentTypeLayoutRow,
+    DotCMSResponse,
+    FieldType
+} from '@dotcms/dotcms-models';
 
 export type DotFieldFilter =
     | 'SHOW_IN_LIST'
@@ -33,5 +39,77 @@ export class DotFieldService {
                 DotCMSAPIResponse<DotCMSContentTypeField[]>
             >(`/api/v3/contenttype/${contentType}/fields/allfields`, { params })
             .pipe(map((x) => x?.entity));
+    }
+
+    /**
+     * Get the field types
+     *
+     * @return {*}  {Observable<FieldType[]>}
+     * @memberof DotFieldService
+     */
+    loadFieldTypes(): Observable<FieldType[]> {
+        return this.#http
+            .get<DotCMSResponse<FieldType[]>>('/api/v1/fieldTypes')
+            .pipe(map((response) => response.entity));
+    }
+
+    /**
+     * Save content type's layout.
+     * @param string contentTypeId Content Type'id
+     * @param DotContentTypeField[] fields fields to add
+     * @returns Observable<DotCMSContentTypeLayoutRow[]>
+     * @memberof DotFieldService
+     */
+    saveFields(
+        contentTypeId: string,
+        fields: DotCMSContentTypeLayoutRow[]
+    ): Observable<DotCMSContentTypeLayoutRow[]> {
+        return this.#http
+            .put<
+                DotCMSResponse<DotCMSContentTypeLayoutRow[]>
+            >(`/api/v3/contenttype/${contentTypeId}/fields/move`, { layout: fields })
+            .pipe(map((response) => response.entity));
+    }
+
+    /**
+     * Delete fields
+     *
+     * @param {string} contentTypeId content types's id that contains the fields
+     * @param {DotCMSContentTypeField[]} fields Fields to delete
+     * @returns {Observable<{ fields: DotCMSContentTypeLayoutRow[]; deletedIds: string[] }>}
+     * @memberof DotFieldService
+     */
+    deleteFields(
+        contentTypeId: string,
+        fields: DotCMSContentTypeField[]
+    ): Observable<{ fields: DotCMSContentTypeLayoutRow[]; deletedIds: string[] }> {
+        return this.#http
+            .request<
+                DotCMSResponse<{ fields: DotCMSContentTypeLayoutRow[]; deletedIds: string[] }>
+            >('DELETE', `/api/v3/contenttype/${contentTypeId}/fields`, {
+                body: {
+                    fieldsID: fields.map((field: DotCMSContentTypeField) => field.id)
+                }
+            })
+            .pipe(map((response) => response.entity));
+    }
+
+    /**
+     * Update a field
+     *
+     * @param {string} contentTypeId content type's id
+     * @param {DotCMSContentTypeField} field field to update
+     * @returns {Observable<DotCMSContentTypeLayoutRow[]>}
+     * @memberof DotFieldService
+     */
+    updateField(
+        contentTypeId: string,
+        field: DotCMSContentTypeField
+    ): Observable<DotCMSContentTypeLayoutRow[]> {
+        return this.#http
+            .put<
+                DotCMSResponse<DotCMSContentTypeLayoutRow[]>
+            >(`/api/v3/contenttype/${contentTypeId}/fields/${field.id}`, { field: field })
+            .pipe(map((response) => response.entity));
     }
 }
