+++ core-web/libs/dotcms-models/src/lib/dot-content-types.model.ts
@@ -656,3 +656,15 @@ export interface DotContentTypePaginationOptions {
     type?: string;
     ensure?: string;
 }
+
+/**
+ * Describes an available field type as returned by the `/api/v1/fieldTypes` endpoint.
+ * Used to render the list of fields that can be added to a content type.
+ */
+export interface FieldType {
+    id: string;
+    label: string;
+    clazz: string;
+    helpText: string;
+    properties: string[];
+}
